*-component -> Actor / Render / Commander

+++ src/config.h
@@ -12,8 +12,8 @@
 using object_id_t = uint32_t; ///< 对象唯一id标识.
 using ref_count_t = uint64_t; ///< 引用计数, 非原子类型. \note 非原子类型.
 
-using tick_t  = uint64_t; ///< 滴答.
-using frame_t = uint64_t; ///< 帧id.
+using tick_t  = uint32_t; ///< 滴答.
+using frame_t = uint32_t; ///< 帧id.
 using float_t = float;    ///< 统一浮点类型.
 using glm::vec2;          ///< 2D向量.
